Update import paths to not have Godeps in it
This is just a temp work around as we had problems importing
cloud init code into our code base.  We ran into two problems.
First, the Docker version cloud-init uses is incompatible with the
version of Docke we use.  Second, the Godeps in the import paths
seemed to confuse `godep save`.  We will circle back and try to
fix whatever we can upstream so that we don't have to maintain a
branch.